+++ crates/playwright/src/protocol/mod.rs
@@ -68,6 +68,7 @@ pub mod wait_for;
 pub mod web_error;
 pub mod web_socket;
 pub mod web_socket_route;
+pub mod web_storage;
 pub mod worker;
 
 pub use accessibility::{Accessibility, AccessibilitySnapshotOptions};
@@ -135,8 +136,8 @@ pub use route::{
     FulfillOptions, FulfillOptionsBuilder, Route, UnrouteBehavior,
 };
 pub use screencast::{
-    ActionPosition, ChapterOptions, OverlayId, Screencast, ScreencastFrame, ScreencastSize,
-    ScreencastStartOptions, ShowActionsOptions, ShowOverlayOptions,
+    ActionCursor, ActionPosition, ChapterOptions, OverlayId, Screencast, ScreencastFrame,
+    ScreencastSize, ScreencastStartOptions, ShowActionsOptions, ShowOverlayOptions,
 };
 pub use screenshot::{Animations, Caret, Scale, ScreenshotClip, ScreenshotOptions, ScreenshotType};
 pub use select_option::SelectOption;
@@ -149,4 +150,5 @@ pub use wait_for::{WaitForOptions, WaitForOptionsBuilder, WaitForState};
 pub use web_error::{WebError, WebErrorLocation};
 pub use web_socket::WebSocket;
 pub use web_socket_route::{WebSocketRoute, WebSocketRouteCloseOptions};
+pub use web_storage::{WebStorage, WebStorageKind};
 pub use worker::Worker;